Refactor doorbird to avoid using events internally
This integration was using events to communicate internally

The doorbird_reset_favorites event is no longer fired
when the clear webhook is called.
