+++ homeassistant/components/doorbird/device.py
@@ -145,3 +145,19 @@ class ConfiguredDoorBird:
             "html5_viewer_url": self._device.html5_viewer_url,
             ATTR_ENTITY_ID: self._event_entity_ids.get(event),
         }
+
+
+async def async_reset_device_favorites(
+    hass: HomeAssistant, door_station: ConfiguredDoorBird
+) -> None:
+    """Handle clearing favorites on device."""
+    await hass.async_add_executor_job(_reset_device_favorites, door_station)
+
+
+def _reset_device_favorites(door_station: ConfiguredDoorBird) -> None:
+    """Handle clearing favorites on device."""
+    # Clear webhooks
+    favorites: dict[str, list[str]] = door_station.device.favorites()
+    for favorite_type, favorite_ids in favorites.items():
+        for favorite_id in favorite_ids:
+            door_station.device.delete_favorite(favorite_type, favorite_id)
