+++ homeassistant/components/doorbird/util.py
@@ -1,5 +1,7 @@
 """DoorBird integration utils."""
 
+from typing import Any
+
 from homeassistant.core import HomeAssistant
 
 from .const import DOMAIN
@@ -7,7 +9,7 @@ from .device import ConfiguredDoorBird
 from .models import DoorBirdData
 
 
-def get_mac_address_from_door_station_info(door_station_info):
+def get_mac_address_from_door_station_info(door_station_info: dict[str, Any]) -> str:
     """Get the mac address depending on the device type."""
     return door_station_info.get("PRIMARY_MAC_ADDR", door_station_info["WIFI_MAC_ADDR"])
